+++ src/main/kotlin/com/github/lorenzomilicia/resourcepartitioninglib/utils/ResourceUtils.kt
@@ -3,7 +3,7 @@ package com.github.lorenzomilicia.resourcepartitioninglib.utils
 import org.springframework.core.io.Resource
 import java.nio.file.Files
 
-internal fun Resource.countLines(): Long =
+internal fun Resource.countLines(): Int =
 	Files.lines(file.toPath()).use {
-		it.count()
+		it.count().toInt()
 	}
